small

+++ code/pairgenz.py
@@ -109,7 +109,7 @@ class PairGen():
             points *= treecat['dcm_mpc']
         self.tree = KDTree(list(points.T))
 
-        self.cat1zshells = np.full(len(self.cat1), 0)
+        self.cat1zshells = np.full(len(self.cat1), 0, dtype='int64')
         self.trees = [self.tree]
         self.treecats = [self.cat2]
